zip handler: eliminate loops for chunk discovery
Zip handler used the initial local file header signature to find
zip chunks. Once the local file header was found the end of
central directory signature was looked for, then the distance between
the two was used to validate the match.

This is not optimal, as there are one local file signature for every
file in a zip file, resulting in a huge amount of signature matches
and also a lot of failed chunk calculation. Worse still, for every match
there are potentially many end of central directory headers to check
against.

The new method directly looks only at the end of central directory
headers, and from that calculates the first local file header
(= start of the chunk).