correctness fix: do not optimize away all chunks that overlap with another
The zip handler has changed the pattern to be matched from start marker
(PK 03 04 - "local file header" - "[" below) to end marker (PK 05 06 -
"end of central directory" - "]" below).
As a side effect some of the extractions have changed. In particular
there was one extraction, where the results had a file with the same
name, but different content (and path).

Let's use X and Y for the files with the same name, but different path,
that were found in the output before and after the zip handler change,
respectively (not at the same time!).
The correct extraction would have both files at the same time.

It was caused by a twin overlap between zip & zlib chunks:
[.X.(..]..)  (..[..).Y.]
where zip start-end is marked with [], and zlib is marked with ().

So the chunks in the input are:
- zip:  [.X.(..]
- zlib:     (..]..)
- zlib: (..[..)
- zip:     [..).Y.]
- (there might be unknown data between the 2 overlapping chunks,
   but it is irrelevant)

Due to the optimization changed, one of the zip files are always
skipped, because there is a zlib chunk, that contain the zip match
patterns in both cases.

----

Some further work is still needed, as by fixing this optimization,
a lot of new false zlib chunks are extracted, producing a lot of noise
in the output.
False zlib chunks are due to zlib having an exceptionally short pattern
(only 2 bytes), and an implicit stream terminator.